+++ src/token.ts
@@ -1463,7 +1463,7 @@ export class MediaWiki {
 					if (lang === 'wiki' || lang === 'wikitext') {
 						lang = 'mediawiki';
 					}
-					if (lang in this) {
+					if (lang in this.config.tagModes && lang in this) {
 						state.extMode = this[lang as 'text/pre']();
 					}
 				} else if (mt[0].trim() === 'ABC') {